requirements.txt
Jinja2
requests
Flask
zhconv
pydantic
tinydb
regex
bbcode
pillow
beautifulsoup4
tenacity
lxml
cchardet